move guard to compiler pass
